add CRUD new tasks and routing. Final project

+++ 17. ToDoList/app.js
@@ -29,6 +29,15 @@ const item3 = new Item({
 
 const defaultItem = [item1, item2, item3]; // Создание пунктов по умолчанию и массива из них
 
+const listSchema = {
+  name: String,
+  items: [itemsSchema]
+} // Создание нового документа для новой маршрута
+
+const List = mongoose.model("List", listSchema); // создание модели по новому документу
+
+
+
 const port = 3000;
 
 let newAdds = [];
@@ -40,7 +49,7 @@ app.set("view engine", "ejs"); // Подключение EJS
 
 app.get("/", function (req, res) {
   Item.find({}, function (err, foundItems) {
-    if (foundItems.length===0) { // Проверка на пустой список
+    if (foundItems.length === 0) { // Проверка на пустой список
       Item.insertMany(defaultItem, function (err) { // Вставка элементов по умолчанию
         console.log(err);
       }); // Добавление массива в бд
@@ -53,19 +62,30 @@ app.get("/", function (req, res) {
 
 app.post("/", function (req, res) {
   const itemName = req.body.newInput; // Захват нового элемента
-  
+  const listName = req.body.listName;
+
   const newItem = new Item({ // Создание нового объекта в бд
     name: itemName,
   });
 
-  newItem.save(); // Сохранение в бд
-  res.redirect("/"); // Отрисовка странциы снова для отображения элементов
+  if (listName === "Today") {
+    newItem.save(); // Сохранение в бд
+    res.redirect("/"); // Отрисовка странциы снова для отображения элементов
+  } else {
+    List.findOne({ name: listName }, function (err, foundList) {
+      foundList.items.push(item);
+      foundList.save();
+      res.redirect("/" + listName);
+    })
+  }
 });
 
 app.post("/delete", function (req, res) {
   const checkItemId = req.body.checkbox; // Захват данных со страницы
+  const listName = req.params.listName;
 
-  Item.findByIdAndRemove(checkItemId, function (err) { // с помощью id удаление из базы 
+  if (listName === "Today") {
+    Item.findByIdAndRemove(checkItemId, function (err) { // с помощью id удаление из базы 
       if (err) {
         console.log(err);
       } else {
@@ -73,11 +93,38 @@ app.post("/delete", function (req, res) {
         res.redirect("/"); // Отрисовка странциы снова для отображения элементов
       }
     });
+  } else {
+    List.findOneAndUpdate({ name: listName }, { $pull: { items: { _id: checkItemId } } }, function (err, foundList) {
+      if (!err) {
+        res.redirect("/" + listName);
+      }
+    });
+  }
 });
 
 // Создание рендера для запроса нового листа для работы
-app.get("/work", function (req, res) {
-  res.render("list", { listTitle: "Work List", newListItems: workAdds });
+app.get("/:customNameList", function (req, res) {
+  const customNameList = _.capitalize(req.params.customNameList);
+  // Поиск уже существующих листов
+  List.findOne({ name: customNameList }, function (err, foundList) {
+    if (!err) {
+      if (!foundList) {
+        // Создание нового листа
+        // Новый список для нового маршрута
+        const list = new List({
+          name: customNameList,
+          items: defaultItem
+        });
+        list.save();
+        res.redirect("/" + customNameList);
+      } else {
+        // Показ найденного листа
+        res.render("list", { listTitle: foundList.name, newListItems: foundList.items })
+      }
+    }
+  })
+
+
 });
 
 app.get("/about", function (req, res) {
